+++ src/main/java/nextstep/courses/domain/session/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public enum SessionStatus {
     READY, RECRUITING, CLOSED
