+++ src/test/java/nextstep/courses/domain/image/SessionImageTest.java
@@ -1,9 +1,8 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.courses.domain.image.SessionImage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
